Changes to BSMFramework: removed declaration of some BJetness variables from EventInfoSelector.cc, Updated Tag SF discriminator in BTagReweight.cc, updated JEC sources in python (MC + DATA)

+++ BSM3G_TNT_Maker/interface/BSM3G_TNT_Maker.h
@@ -95,7 +95,7 @@ class BSM3G_TNT_Maker : public edm::EDAnalyzer {
   /////
   TFile* file;
   TTree* evtree_;
-  int eventnum;
+  long long eventnum;
   int eventnumnegative;
   TTree* tree_;
   const size_t MaxN;
